mat_op

Assemble several matrices of same dimensions by performing by case operation. For instance, with kind.of.operation = "+", the function will add the value of all the case 1 (row1 & column1) of the matrices and put it in the case 1 of a new matrix M, according to the formula

$$c_{ij} = \sum_{k=1}^{k=z}c_{ijk}$$ c: case

i: row number

j: column number

k: matrix number

z: number of matrices

mat_op(mat.list, kind.of.operation = "+")

Arguments

mat.list

List of matrices.

kind.of.operation

Either "+" (by case addition), "-" (by case subtraction) or "*" (by case multiplication).

Value

The assembled matrix, with row and/or column names only if all the matrices have identical row/column names.

Details

REQUIRED PACKAGES

cuteDev

cuteTool

REQUIRED FUNCTIONS FROM THE cute PACKAGE

arg_check()

comp_2d()

Examples

mat1 = matrix(c(1,1,1,2,1,5,9,8), ncol = 2) ; 
mat2 = matrix(c(1,1,1,2,1,5,9,NA), ncol = 2) ; 
mat_op(mat.list = list(mat1, mat2), kind.of.operation = "+")
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2   10
#> [3,]    2   18
#> [4,]    4   NA

mat1 = matrix(c(1,1,1,2,1,5,9,8), ncol = 2, dimnames = list(LETTERS[1:4], letters[1:2])) ; 
mat2 = matrix(c(1,1,1,2,1,5,9,NA), ncol = 2, dimnames = list(LETTERS[1:4], letters[1:2])) ; 
mat_op(mat.list = list(mat1, mat2), kind.of.operation = "*")
#>   a  b
#> A 1  1
#> B 1 25
#> C 1 81
#> D 4 NA

mat1 = matrix(c(1,1,1,2,1,5,9,8), ncol = 2, dimnames = list(LETTERS[1:4], c(NA, NA))) ; 
mat2 = matrix(c(1,1,1,2,1,5,9,NA), ncol = 2, dimnames = list(LETTERS[1:4], letters[1:2])) ; 
mat_op(mat.list = list(mat1, mat2), kind.of.operation = "-")
#>   [,1] [,2]
#> A    0    0
#> B    0    0
#> C    0    0
#> D    0   NA

mat1 = matrix(c(1,1,1,2,1,5,9,8), ncol = 2, dimnames = list(c("A1", "A2", "A3", "A4"), letters[1:2])) ; 
mat2 = matrix(c(1,1,1,2,1,5,9,NA), ncol = 2, dimnames = list(LETTERS[1:4], letters[1:2])) ; 
mat3 = matrix(c(1,1,1,2,1,5,9,NA), ncol = 2, dimnames = list(LETTERS[1:4], letters[1:2])) ; 
mat_op(mat.list = list(mat1, mat2, mat3), kind.of.operation = "+")
#>      a  b
#> [1,] 3  3
#> [2,] 3 15
#> [3,] 3 27
#> [4,] 6 NA